Regulisan problem sa funkcijama i tipovima

+++ trans/rekurzija/rekiter.c
@@ -11,7 +11,7 @@ void viter() {
   putchar('\n');
 }
 
-int itenr(double a, int rez) {
+int itenr(long a, int rez) {
   if (a <= 0)
     return rez;
   else
@@ -19,7 +19,7 @@ int itenr(double a, int rez) {
 }
 
 // duboki polozaj poziva
-int itenr2(double a, int rez) {
+int itenr2(long a, int rez) {
   if (a <= 0) {
     rez--;
     {
